Test setup fix

+++ test/setup/clearScratch.ts
@@ -3,5 +3,7 @@ const path = require('path');
 
 const SCRATCH_PATH = './scratch';
 
-const files = fs.readdirSync(SCRATCH_PATH);
-files.forEach((file: string) => fs.rmSync(path.join(SCRATCH_PATH, file)));
+if (fs.existsSync(SCRATCH_PATH)) {
+  const files = fs.readdirSync(SCRATCH_PATH);
+  files.forEach((file: string) => fs.rmSync(path.join(SCRATCH_PATH, file)));
+}
